Move demo components to components subdirectory
