+++ apps/playground/app/ui/button/page.tsx
@@ -1,6 +1,6 @@
 import { Button } from "@giselle-internal/ui/button";
-import { DemoSection } from "../demo-section";
-import { UiPage } from "../ui-page";
+import { DemoSection } from "../components/demo-section";
+import { UiPage } from "../components/ui-page";
 
 export default function () {
 	return (
