+++ apps/playground/app/ui/dropdown-menu/page.tsx
@@ -2,8 +2,8 @@
 
 import { Button } from "@giselle-internal/ui/button";
 import { DropdownMenu } from "@giselle-internal/ui/dropdown-menu";
-import { DemoSection } from "../demo-section";
-import { UiPage } from "../ui-page";
+import { DemoSection } from "../components/demo-section";
+import { UiPage } from "../components/ui-page";
 
 export default function () {
 	return (
