+++ apps/playground/app/ui/input/page.tsx
@@ -1,6 +1,6 @@
 import { Input } from "@giselle-internal/ui/input";
-import { DemoSection } from "../demo-section";
-import { UiPage } from "../ui-page";
+import { DemoSection } from "../components/demo-section";
+import { UiPage } from "../components/ui-page";
 
 export default function () {
 	return (
