+++ apps/playground/app/ui/popover/page.tsx
@@ -2,8 +2,8 @@
 
 import { Button } from "@giselle-internal/ui/button";
 import { Popover } from "@giselle-internal/ui/popover";
-import { DemoSection } from "../demo-section";
-import { UiPage } from "../ui-page";
+import { DemoSection } from "../components/demo-section";
+import { UiPage } from "../components/ui-page";
 
 export default function () {
 	return (
