+++ apps/playground/app/ui/select/page.tsx
@@ -1,6 +1,6 @@
 import { Select } from "@giselle-internal/ui/select";
-import { DemoSection } from "../demo-section";
-import { UiPage } from "../ui-page";
+import { DemoSection } from "../components/demo-section";
+import { UiPage } from "../components/ui-page";
 
 export default function () {
 	return (
