+++ apps/playground/app/ui/table/page.tsx
@@ -8,8 +8,8 @@ import {
 	TableHeader,
 	TableRow,
 } from "@giselle-internal/ui/table";
-import { DemoSection } from "../demo-section";
-import { UiPage } from "../ui-page";
+import { DemoSection } from "../components/demo-section";
+import { UiPage } from "../components/ui-page";
 
 export default function () {
 	return (
