+++ apps/playground/app/ui/toggle/page.tsx
@@ -1,6 +1,6 @@
 import { Toggle } from "@giselle-internal/ui/toggle";
-import { DemoSection } from "../demo-section";
-import { UiPage } from "../ui-page";
+import { DemoSection } from "../components/demo-section";
+import { UiPage } from "../components/ui-page";
 
 export default function () {
 	return (
